Made Derivative Tests, fixed error in the derivative method and fixed Term
I made derivative tests to the program. I also fixed an issue with the derivative method, since there was going to be an error when it read the first part of the equation if a negative existed. I also fixed the term class since I used a break statement, so I was able to get ride of it.