Derivative buttons added, few exceptions added
With the derivative test working, I added the buttons on the calculator. There are also a few exceptions when the user tries to press equals or derivative button at the beginning when it first displays the message. I also added a method in the interface, and switched a piece of the if statement in the derivative method since it would go index out of bounds if I did not.