Documentation, Term updates, and more tests
I added more tests into the program to calculate the evaluate method. I also fixed an issue with term, since it wouldn't calculate the derivative if a single "x" character existed. There is also a lot of documentation added.
added a method in the interface, and switched a piece of the if statement in the derivative method since it would go index out of bounds if I did not.