UML update, documentation update.
UML is updated to include the mandatory classes in the UML guide, along with the CalculatorModel. More documentation has been added in several areas.
ative if a single "x" character existed. There is also a lot of documentation added.
added a method in the interface, and switched a piece of the if statement in the derivative method since it would go index out of bounds if I did not.